exercicios de python

+++ estruturas_condicionais_if-else-simples-iii.py
@@ -1,6 +1,6 @@
 #Simula um radar de velocidade, informando se o veículo ultrapassou a máxima permitida e quanto ficou a multa
 print('#######RADAR DE VELOCIDADE#######')
-vmax = int(80+(20/100*80))
+vmax = float(80+(20/100*80))
 multa = float(7.00)
 vdetectada = int(input('Velocidade do veículo: '))
 if vdetectada > vmax:
