+++ estruturas_condicionais_if-else-simples-viii.py
@@ -1,7 +1,7 @@
 #Informa se três retas podem ou não formar um triângulo
-l1 = int(input('Informe o comprimento da 1a reta: '))
-l2 = int(input('Informe o comprimento da 2a reta: '))
-l3 = int(input('Informe o comprimento da 3a reta: '))
+l1 = float(input('Informe o comprimento da 1a reta: '))
+l2 = float(input('Informe o comprimento da 2a reta: '))
+l3 = float(input('Informe o comprimento da 3a reta: '))
 if l1 > abs(l2 - l3) and l1 < l2 + l3 or l2 > abs(l1 - l3) and l2 < l1 + l3 or l3 > abs(l1 - l2) and l3 < l1 + l2:
     print('As retas podem formar um triângulo.')
 else:
